# Rearrange halo search