Performance improvement (checked on GPU) and simplification of particle halo search. Should enable better reuse for contact ghost neighbors