@streeve 2023-09-23
Speedups using the current crack branching and elastic wave with double the size per dimension. "Other" speedup is possibly unrelated

Previous:
| Ranks | Particles | Time | T_Force | T_Comm | T_Int | T_Other | T_Init |
| -------- | ------- |------- |------- |------- |------- |------- |------- |
| 8 | 512000 | 9.93 | 2.79 | 4.82 | 0.04 | 2.14 | 0.12 | PERFORMANCE (crack) | 
| 8 | 531441 | 2.45 | 0.97 | 1.19 | 0.00 | 0.08 | 0.20 | PERFORMANCE (elastic) | 

New:
| Ranks | Particles | Time | T_Force | T_Comm | T_Int | T_Other | T_Init |
| -------- | ------- |------- |------- |------- |------- |------- |------- |
8 | 512000 | 7.94 | 2.80 | 3.69 | 0.04 | 1.28 | 0.12 | PERFORMANCE
8 | 531441 | 2.24 | 0.97 | 0.96 | 0.00 | 0.11 | 0.20 | PERFORMANCE

Speedup:
| Total | Force | Comm | Int | Other | Init |
| -------- | ------- |------- |------- |------- |------- |
| 1.25 | 0.99 | 1.30 | 1.0 | 1.67 | 1.0 |
| 1.09  | 1.0  |1.24  |1.0  |0.72  |1.0 |